Add android build capability to program and make basic login view
