+++ application/bus_me/android/app/build.gradle
@@ -27,7 +27,7 @@ apply from: "$flutterRoot/packages/flutter_tools/gradle/flutter.gradle"
 
 android {
     namespace "com.busme.bus_me"
-    compileSdkVersion 20
+    compileSdkVersion 33
     ndkVersion flutter.ndkVersion
 
     compileOptions {
