+++ application/bus_me/android/app/src/debug/AndroidManifest.xml
@@ -1,7 +1,15 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.busme.bus_me">
     <!-- The INTERNET permission is required for development. Specifically,
          the Flutter tool needs it to communicate with the running application
          to allow setting breakpoints, to provide hot reload, etc.
     -->
     <uses-permission android:name="android.permission.INTERNET"/>
+
+    <application>
+        <activity android:name="com.busme.bus_me.MainActivity" android:exported = "true"></activity>
+    </application>
+
+
+
+
 </manifest>
