+++ application/bus_me/android/app/src/profile/AndroidManifest.xml
@@ -4,4 +4,8 @@
          to allow setting breakpoints, to provide hot reload, etc.
     -->
     <uses-permission android:name="android.permission.INTERNET"/>
+
+    <application>
+    </application>
+
 </manifest>
