+++ application/bus_me/lib/models/pin_pill_info.dart
@@ -1,12 +1,13 @@
 import 'dart:ui';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
+
 class PinInformation {
-  String pinPath;
-  String avatarPath;
-  LatLng location;
-  String locationName;
-  Color labelColor;
+  String? pinPath;
+  String? avatarPath;
+  LatLng? location;
+  String? locationName;
+  Color? labelColor;
 
   PinInformation({this.pinPath, this.avatarPath, this.location, this.locationName, this.labelColor});
 }
