+++ application/bus_me/macos/Flutter/ephemeral/Flutter-Generated.xcconfig
@@ -1,6 +1,6 @@
 // This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=/Applications/flutter
-FLUTTER_APPLICATION_PATH=/Users/anushkar/work/aut/Bus_App/bus_app
+FLUTTER_ROOT=C:\Users\jgreen\Documents\flutter
+FLUTTER_APPLICATION_PATH=C:\Users\jgreen\Documents\Git\BusME\application\bus_me
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_BUILD_DIR=build
 FLUTTER_BUILD_NAME=1.0.0
